+++ src/main/java/duke/Duke.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.command.Command;
+import duke.command.Parser;
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
 public class Duke {
 
     private Storage storage;
