+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
+public abstract class Command {
+
+    public void execute(TaskList tasks, Ui ui, Storage storage) {}
+
+    public abstract boolean isExit();
+}
