+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 public class MarkCommand extends Command {
 
     private int index;
