+++ src/main/java/duke/command/Parser.java
@@ -1,4 +1,4 @@
-import java.text.NumberFormat;
+package duke.command;
 
 public class Parser {
